Change battle system
Each region now accounts for both defending and invading forces.
When a region has an invading force, a battle is generated.
In a mission with battle(s), tactical column movements are paused.
When the mission finishes, results are extracted, damages applied to
travelling tanks and to battling tanks.
Region tank contents are updated and battles are resolved, accounting for
bonuses and damages.