Move module-level variable to function, in an attempt to fix a potential race condition
The problem resides in an NRE pointing at

  match valuedObjects.TryGetValue(entry.ObjectType.ToLowerInvariant()) with

The exception is non-deterministic, which eliminates entry.ObjectType being
null due to a failed parsing.

My guess is that the async execution of the asyncSeq closure containing
that line might race with the creation of valuedObjects, which I think
happens when the module is first accessed.