Stop and start commenter from plugin
The old method of watching state.xml was a bad idea: the commentator
would restart while the mission files are being generated and resaved.
Meanwhile, players who are still playing the old mission would be handled
as if they were in the new mission (that isn't ready yet), which would
lead to weird behaviour.