Redesign WarState to be mutable
The mutability helps write more concise code to update the state.
The potential risks of side effects shouldn't be a problem, as we typically
have only one war state object, which is modified step by step in a linear
pipeline.

Also moved the status and descriptions of buildings from the objects that
use them, to the top level, and use references through identifiers to
express usage relatiships instead.