Add Hurricane MkII to plane DB
The plane has not yet been added to any planeset, although it's planned
in e.g. the Stalingrad planeset. The reason is that the mechanism
to exclude some of the weapon mods is not yet implemented, and it would
not be fair to give all weapon mods in all scenarios.