Work around bug causing too many captured pilots
The fast-forward feature is not deterministic, leading to different
outcomes in region captures, leading to pilots taking off and landing
at airfields now owned by the enemy.

Work around makes the "land in enemy territory" more strict by comparing
owner of start and end airfields. If they are the same, the pilot is
not considered captured.