Clarify the role of PlayerFlightState
Prefer "player mission" instead of "player flight", as "flight" leads
to confusion with the flight record.

Replace "destroyed" by "ejected". A plane that is consider destroyed
might still be flyable (e.g. dead engine), and landing a destroyed plane
should be possible and properly accounted for.
This is unlike a pilot ejection, which should be considered as ending
possibilities for that pilot to achieve anything.