Rewrite allOnOuterSide condition
Fix earlier incorrect "fix" for border condition when cross product is
null. The condition was actually checking if some vertex was inside,
and therefore should not have been strict.

Rewrote to be less confusing and check if all vertices are outside, as
the name of the variable indicates.

Also skip degenerate edges, instead of falsifying the condition.